cmds: Enable vmem translations in `memmap`
