+++ src/cmds/hardware/memmap/Memmap.my
@@ -24,7 +24,7 @@ package embox.cmd
 module memmap {
 	source "memmap.c"
 
-	depends embox.compat.libc.all
-	depends embox.compat.posix.LibPosix
-	depends embox.framework.LibFramework
+	depends embox.mem.marea_header
+	depends embox.mem.mmap_header
+	depends embox.mem.vmem_api
 }
